@@ -25,11 +25,11 @@
     "release:integrity:verify": "cd releases/ravelinjs-$npm_package_version; sed s/^sha384-// integrity | shasum -c",
     "release:tarball": "cd releases; tar -czf ravelinjs-$npm_package_version.tar.gz ravelinjs-$npm_package_version",
     "test": "BROWSERS=ie,8 LIMIT=1 LOG_LEVEL=error PARALLEL=4 run-s build test:unit test:integration",
-    "test:e2e": "BROWSERS=chrome,latest LIMIT=1 npm run test:integration -- --spec test/encrypt/encrypt.spec.js",
+    "test:e2e": "BROWSERS=chrome,latest LIMIT=1 npm run test:integration -- --spec test/encrypt/encrypt.spec.mjs",
     "test:unit": "karma start ./test/karma.conf.js --single-run",
     "test:unit:watch": "karma start ./test/karma.conf.js --auto",
-    "test:integration": "wdio ./test/wdio.conf.js",
-    "test:integration:quickunit": "env BROWSERS=firefox,latest LIMIT=1 wdio ./test/wdio.conf.js --spec test/unit/unit.spec.js",
+    "test:integration": "wdio ./test/wdio.conf.mjs",
+    "test:integration:quickunit": "env BROWSERS=firefox,latest LIMIT=1 wdio ./test/wdio.conf.mjs --spec test/unit/unit.spec.mjs",
     "update": "ncu --doctor -u -x @wdio/*",
     "md": "remark . -o",
     "md:watch": "remark . -o --watch",
@@ -40,18 +40,17 @@
     "@rollup/plugin-commonjs": "20.0.0",
     "@rollup/plugin-node-resolve": "13.0.4",
     "@rollup/plugin-replace": "3.0.0",
-    "@wdio/browserstack-service": "6.12.1",
-    "@wdio/cli": "6.12.1",
-    "@wdio/junit-reporter": "6.11.0",
-    "@wdio/local-runner": "6.12.1",
-    "@wdio/logger": "6.10.10",
-    "@wdio/mocha-framework": "6.11.0",
-    "@wdio/spec-reporter": "6.11.0",
-    "@wdio/sync": "6.11.0",
+    "@wdio/browserstack-service": "https://github.com/unravelin/webdriverio/releases/download/v8.1.3-severe1/wdio-browserstack-service-8.1.3.tgz",
+    "@wdio/cli": "8.1.3",
+    "@wdio/dot-reporter": "8.3.0",
+    "@wdio/local-runner": "8.1.3",
+    "@wdio/logger": "8.1.0",
+    "@wdio/mocha-framework": "8.1.2",
     "build-url": "6.0.1",
     "chai": "4.3.4",
     "cors": "2.8.5",
     "deep-diff": "1.0.2",
+    "deepmerge": "4.2.2",
     "express": "4.17.1",
     "jshint": "2.13.1",
     "karma": "6.3.4",
@@ -89,8 +88,7 @@
       }
     }
   },
-  "dependencies": {},
   "volta": {
-    "node": "10.24.1"
+    "node": "18.13.0"
   }
 }
