@@ -1,8 +1,8 @@
-const { navigate, hasTitle, hasElement } = require('../common.spec.js');
+import { navigate, hasTitle, hasElement } from '../common.spec.mjs';
 
 describe('ravelinjs unit tests', function() {
-  it('passes', function() {
-    navigate(browser, {
+  it('passes', async function() {
+    await navigate(browser, {
       attempts: 3,
       url: '/unit/',
       tests: [
@@ -15,13 +15,13 @@ describe('ravelinjs unit tests', function() {
     });
 
     // See if we got any errors.
-    var stats = $('#mocha-stats').getText().replace(/(\d)([fd])/g, '$1, $2');
+    const stats = (await $('#mocha-stats').getText()).replace(/(\d)([fd])/g, '$1, $2');
     if (stats.indexOf('failures: 0,') !== -1) {
       return;
     }
 
     // Report back errors.
-    var failures = $('#completed pre.error').getText();
+    const failures = await $('#completed pre.error').getText();
     throw new Error('Stats: ' + stats + '. ' + failures);
   });
 });
